Add smoketest for backend dummy and fix quirks in testing
The smoketests wait for elosd to open a tcp port to know that it's
running. So the tcp plugin needs to be installed and  loaded.
And the scanner_manager_legacy crashes when no legacy scanner plugin is
configured and in the right path. So that needs to be done as well.

Signed-off-by: Friedrich Schwedler <fschwedler@emlix.com>